requirements.txt
evaluate
transformers
accelerate
datasets==2.10.1
SQLAlchemy==1.4.45
geopandas==0.12.2
h11==0.14.0
h5py==3.8.0
huggingface-hub==0.12.0
hydra-core==1.1.0rc1
imageio==2.0.0
keras==2.11.0
matplotlib==3.6.3
numpy==1.21.5
omegaconf==2.1.0rc1
opencv-python
packaging==23.0
pandas==1.5.3
Pillow==9.2.0
plotly==5.13.0
promise==2.3
protobuf==3.19.6
psutil==5.9.4
ptvsd==4.3.2
py==1.11.0
pyarrow==11.0.0
pytorch-lightning==1.3.8
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==5.1.2
rasterio==1.3.6
ray==2.2.0
requests==2.28.2
responses==0.18.0
retry==0.9.2
scipy==1.9.1
sympy==1.11.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.11.0
tensorflow-addons==0.20.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.13.0
tqdm==4.64.1
einops==0.6.1
wandb==0.15.5
scikit-learn
